[기능] 상점 UI 닫기 및 아이템 구매 (코인 소비) 추가

+++ Assets/Scripts/Contents/UI/Shop UI/ShopItem.cs
@@ -10,9 +10,9 @@ public class ShopItem : MonoBehaviour
     [SerializeField] TextMeshProUGUI _itemCostText;
 
     public int ItemId { get; private set; }
-    public float ItemCost { get; private set; }
+    public int ItemCost { get; private set; }
 
-    public void Init(int itemId,float itemCost)
+    public void Init(int itemId,int itemCost)
     {
         ItemId = itemId;
         ItemCost = itemCost;
